Introduce RequestBodyHandler
Before this change, there was only a handler to reject too large
requests which was implemented in BodySizeLimitHandler. But now, there
is also the possibility to check if the request body contains the
desired MediaType. These two functions are now combined in
RequestBodyHandler.

To increase the security, this handler does not just parse the
Content-Type header of the request. The handler really parses the byte
stream with the help of Tika [1].

[1] https://tika.apache.org/

Change-Id: If968d95242fc7d5feeeee5a776bccc8560a188e2